requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/test.txt requirements/base.in requirements/test.in
#
attrs==17.4.0             # via pytest
certifi==2018.1.18        # via requests
chardet==3.0.4            # via requests
click==6.7                # via drf-openapi
coreapi==2.3.3            # via django-rest-swagger, openapi-codec
coreschema==0.0.4         # via coreapi
coverage==4.5.1           # via pytest-cov
django-rest-swagger==2.1.2  # via drf-openapi
djangorestframework==3.7.7
drf-openapi==1.3.0
funcsigs==1.0.2           # via pytest
idna==2.6                 # via requests
itypes==1.1.0             # via coreapi
jinja2==2.10              # via coreschema
markupsafe==1.0           # via jinja2
mysqlclient==1.3.12
openapi-codec==1.3.2      # via django-rest-swagger
pluggy==0.6.0             # via pytest
py==1.5.2                 # via pytest
pytest-cov==2.5.1
pytest-django==3.1.2
pytest==3.4.1             # via pytest-cov, pytest-django
pytz==2018.3              # via django
requests==2.18.4
simplejson==3.13.2        # via django-rest-swagger
six==1.11.0               # via pytest
uritemplate==3.0.0        # via coreapi
urllib3==1.22             # via requests
